Rename api-schema to apiSchema
