+++ .prettierignore
@@ -1,1 +1,1 @@
-api-schema.ts
+apiSchema.ts
